inotify experimental

+++ check.pl
@@ -3,26 +3,34 @@ use warnings;
 use strict;
 use POSIX;
 use config qw(config_load_all);
+use Linux::Inotify2;
 my $pid;
-if($pid=fork){
-	die "Can't fork: $!\n" if $pid<0;
-	exit 0;
+my $nodaemon = '';
+while(shift){
+	$nodaemon = 1 if($_ eq "-D");
 }
-POSIX::setsid or die "setsid: $!\n";
-if($pid=fork){
-	die "Can't fork: $!\n" if $pid<0;
-	my $pidfile;
-	if(!(open $pidfile, '>', "/home/fss/.checker.pid")){
-		kill 9, $pid;
-		die "Can't open pid file\n";
+if(!$nodaemon){
+	print "Forking to background.\n";
+	if($pid=fork){
+		die "Can't fork: $!\n" if $pid<0;
+		exit 0;
 	}
-	print $pidfile $pid;
-	close $pidfile;
-	exit 0;
-}
-umask 0;
-foreach (0 .. (POSIX::sysconf (&POSIX::_SC_OPEN_MAX) || 1024))
+	POSIX::setsid or die "setsid: $!\n";
+	if($pid=fork){
+		die "Can't fork: $!\n" if $pid<0;
+		my $pidfile;
+		if(!(open $pidfile, '>', "/home/fss/.checker.pid")){
+			kill 9, $pid;
+			die "Can't open pid file\n";
+		}
+		print $pidfile $pid;
+		close $pidfile;
+		exit 0;
+	}
+	umask 0;
+	foreach (0 .. (POSIX::sysconf (&POSIX::_SC_OPEN_MAX) || 1024))
 	{ POSIX::close $_ }
+}
 open (STDIN, "</dev/null");
 open (STDERR, ">&STDOUT");
 chdir "/";
@@ -30,6 +38,10 @@ my $base_dir="/fsdata/site/mirror/";
 my $status_dir=$base_dir.".status/";
 open (STDOUT, ">>${status_dir}.checker.log");
 opendir my $sdh, $status_dir or die "Failed to open $status_dir\n";
+my $inotify = new Linux::Inotify2;
+for my $tmp (grep {!/\./ && -d "$status_dir/$_"} readdir($sdh)){
+	$inotify->watch($tmp, IN_ALL_EVENTS);
+}
 opendir my $dh, $base_dir or die "Failed to open $base_dir\n";
 while(1){eval{
 	rewinddir $dh;
@@ -120,5 +132,20 @@ while(1){eval{
 	}else{
 		print "mirror_status.json updated at ", time, "\n";
 	}
-	sleep 5*60;
+	my @events = $inotify->read;
+	unless (@events > 0) {
+		for my $tmp (@events){
+			if($tmp->name eq $status_dir){
+				if($tmp->mask | IN_CREATE){
+					if( -d $status_dir.$tmp->name ){
+						$inotify->watch($status_dir.$tmp->name, IN_ALL_EVENTS);
+					}
+				}elsif($tmp->mask | IN_DELETE_SELF){
+					die "WTF are you doing? Nooooooooooooooooooo";
+				}
+			}elsif($tmp->mask | IN_DELETE_SELF){
+				$tmp->w->cancel();
+			}
+		}
+	}
 }
